Changed video source
I decided to change the video sources from a YT video to local one mainly because its quicker to load and you can easily hide controls and not worry about cards showing. Plus it loops the s2 video much smoother. The s1 video still has the glitch at the end and I can't upload the flv one cos it's not a compatible type, and converting it to mp4 cuts a second off for some reason, and just changing the extension to mp4 doesn't work either.