Separate Browser versions & border
I made one that works in Firefox. It was just a simple change to the manifest file. But it does work in Firefox.
I also tried adding a border around the video in Chrome so it looks like Mr Smith's screen.
I added the same code in the Firefox version but that doesn't seem to work.
Edge still doesn't work. It is the Legacy one. Video seems to play in the html file in Edge, just not in the extension popup.